requirements.txt
matplotlib
numpy
tensorflow
Pillow
keras
h5py